requirements.txt
Django==5.2.7
django-model-utils==5.0.0
django-thumbnails==0.8.0
django-redis==6.0.0

# API
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Storage Stuff
psycopg2-binary==2.9.11